Added new configure option:  --enable-anasazi-experimental Currently this option will enable the compilation of the Jacobi-Davidson code that Marzio is working on.
Fixed MyMultiVec and MyOperator classes in the MVOPComplex directory.  Removed
specialized interface to the MultiVecTraits/OperatorTraits class.
This specialization is not necessary since there is a default interface for any
class that inherits from Anasazi::MultiVec/Operator.  It was also confusing the
compiler because MyMultiVec/Operator is templated on ScalarType and the specialization
used specifically "double".  This was the source of the Anasazi make failures
in the test-harness for the last week.

Fixed the BlockJacobiDavidson example to use the typedefs set at the top of the
file.  Also fixed the ProjectedEigenproblem class so that is uses the MultiVecTraits
class to interface with any multivectors from Anasazi.

Put $(TEUCHOS_LIBS) back in the test library definitions for Anasazi.  Anasazi
is strictly dependent upon Teuchos and, if Thyra is not configured, then Anasazi's
examples and tests will fail to build.