Processor split code in recursive bisection now checks if the #procs in left or right processor group is prime; if they are, #procs is reduced by one hoping to yield a better decomposition (at least it will be even; so it will have 2x? decomposition).
This is hard to test especially with when a similar code exist at the
beginning of the partitioning; I've tested by reducing the
"SMALL_PRIME" constant to 3 and setting np=10 and np=9. np=10 makes
the left partition prime (5) and when it is reduced to 4, second will
have 6 procs. np=9 makes right partition prime (5) and will force to
leave one processor idle.

More testing will be useful but I think the code is working fine :)