For routines TREVC and GEES:
Added additional implementations to deal with inconsistencies between real
and complex scalar types.

Two version exist for each scalar type:
* one which is unique for the scalar type, with respect to the signature
of the SELECT function, the existence of RWORK, and the representation of
eigenvalues (W vs. WR and WI)
* one which is generic and can be used for any scalar type, which contains
represents eigenvalues by WR and WI, and passes both WORK and RWORK, and
does not enable the SELECT functionality for sorting/selecting.

The latter was necessary for generic programming using GEES and TREVC.