----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	src/Kokkos_OskiMatrix.hpp
 ----------------------------------------------------------------------
Added the ability to specify the matrix properties recognized by OSKI
as listed on page 19 of the OSKI User Guide in Table 3.

Instead of a Teuchos parameter list, I used 5 optional parameters for now.
I have not determined if key/value pairs would be better or not as we
would have to determine keys for each group of values, which might
needlessly complicate things.

Tests pass, but should be expanded to exercise the new capability
more fully.