Added an optional bool argument to a couple of Epetra_LinearProblem constructors and 'Set' methods, so that Epetra_LinearProblem can, at the user's direction, take ownership of the matrix and/or vector objects meaning that Epetra_LinearProblem can delete them when it is destroyed.
The bool argument is optional, and defaults to false, so no existing
code needs to be modified, and no behavior is changed.

This allows for factory-style code that creates a matrix and vectors
and uses them to create a Epetra_LinearProblem object which is passed out
to a calling scope, without requiring that the Epetra_LinearProblem's
matrix/vector attributes be separately destroyed at the appropriate
time.

Verified that all tests pass.