----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	configure configure.ac
 ----------------------------------------------------------------------
Bug 1415.

Modified the math library error message because when configure dies
looking for pow, it is almost always because of a problem not related
to the math library.  The message now reads:

configure: error: Error encountered when checking for math library.  ***Almost
always this error is caused by a typo in the list of configure arguments
that is not related to the math library.  Look especially for missing
quotes or hyphens in any value related to compiler flags or libraries.
For an error message that may assist in finding the problem grep for
"checking for library containing pow" in the file
packages/teuchos/config.log and read the next several lines.

Mike Heroux suggested that this change be made to Teuchos only because
Teuchos is required for almost every other Trilinos package.

Bug 1906.

Made minor modifications to the compiler selection lists.  Removed
gcc from the c++ list (unsure why it was there).  Also added gfortran,
the replacement for g77, and f95 to the fortran list.

I tested these changed on my local Linux box.