Several restructuring activities are committed here.
Isorropia::Partitioner has been turned into an abstract base class,
and there is now a concrete implementation of it, called
Isorropia::EpetraPartitioner.

Isorropia::Redistributor has not yet been made into an abstract class.

Isorropia::Epetra_Utils now has a function which accepts a Partitioner,
and returns a target-map.

The create_balanced_map functions have been renamed to 'repartition',
and now produce import/export sets and no longer return a map. They
are now more suitable for use as the kernel for the Partitioner class.

More restructuring is still to be done, but this is a good commit point,
because the tests are working again (they were all broken temporarily
during the restructuring).