----------------------------------------------------------------------  Committing in .
 Modified Files:
 	CSEDevModelPaper.tex
 ----------------------------------------------------------------------
Cleaned up wording in a few spots.  Added a couple missing names to the
list of Trilinos developers.  Removed Mike H's name from the list
(because he is an author) and changed the wording to read "our fellow
Trilinos developers".  If someone doesn't like this, it can certainly
be changed back.