Address bugs 1945 and 2273.
1945 concerns Epetra_LinearProblem not querying A.UseTranspose() to
decide whether to do left- or right-scaling. Now, the LeftScale()
methods performs left-scaling if A.UseTranspose is false, else it
performs right-scaling. Conversely for the RightScale() method.

2273 concerns adding a 'IsOneToOne()' query to Epetra_BlockMap.
The Epetra_BlockMap constructors now call a function which creates
a directory on the BlockMapData object, if it doesn't already exist,
and then queries the directory to determine whether all GIDs are
uniquely owned (appear on only one processor).

All tests run and pass.